fastapi/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
requests
bs4
fake_useragent
lxml